Added game pieces and basic movement functionality
The game now automatically loads in the 32 board pieces and lets the player be a certain color based on what the global player variable is set to. Pieces can move in every direction besides diagonally backward to add a more interesting limitation to movement options and encourage forward movement. Highlighting is done automatically. Pieces can move on top of each other but cannot be captured yet.